Polishing after self-review
